+++ libs/vkd3d/vkd3d_private.h
@@ -2409,7 +2409,7 @@ struct d3d12_pipeline_state
 #ifdef VKD3D_ENABLE_PROFILING
     struct
     {
-        size_t pso_entry_index;
+        uint32_t pso_entry_index;
     } timestamp_profiler;
 #endif
 };
@@ -3492,7 +3492,7 @@ struct d3d12_command_list
     struct
     {
         struct d3d12_pipeline_state *active_timestamp_state;
-        size_t timestamp_index;
+        uint32_t timestamp_index;
 
         struct vkd3d_timestamp_profiler_submitted_work *work;
         size_t work_count;
